BugFix (no tmp directory causes exception)

+++ Editor/EditorWindow.cs
@@ -1172,7 +1172,8 @@ namespace ARdevKit
                 e.Cancel = true;
                 return;
             }
-            Directory.Delete("tmp", true);
+            if (Directory.Exists("tmp"))
+                Directory.Delete("tmp", true);
         }
 
         /// <summary>
